refac: Move query functions from vuln module to the db_interface module #2462
